Reuse ProcessRowView with valueViewBuilder
